A bunch of minor tweaks
I had the version number on the about screen hard-coded to 1.2.0. Doh! Fixed it to use the app bundle short version instead.
Changed color of Reset button text. In dark mode, the text what light grey and that didn’t look good on the yellow background.
The tap area on the reset button was just the label text, not the entire button. Fixed it so you could tap the entire button.